Hand object
Player can successfully draw a card from the deck to their hand. This removes a card from the deck. Additionally, A To_String was added to card.

Current issue is that active_player cannot switch between players. This is likely due to active_player not properly being assigned as a reference.